refactor for helper
